+++ sdks/outpost-python/RELEASES.md
@@ -139,3 +139,13 @@ Based on:
 - [python v1.1.0] sdks/outpost-python
 ### Releases
 - [PyPI v1.1.0] https://pypi.org/project/outpost_sdk/1.1.0 - sdks/outpost-python
+
+## 2026-05-01 17:33:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.10 (2.881.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.2.0] sdks/outpost-python
+### Releases
+- [PyPI v1.2.0] https://pypi.org/project/outpost_sdk/1.2.0 - sdks/outpost-python